Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int a, b;
+    float a, b;
     char opr;
-    scanf("%d %d %c",&a, &b,&opr);
+    scanf("%f %f %c",&a, &b,&opr);
     switch(opr){
         case '+':
         printf("%d", (a + b));
